Revert "TACO-176 create amazon token from HEM (#2163)"
This reverts commit 3b734b11380f891383c0d9c95e36bdcfd00a094b.
