+++ src/platforms/shared/modes/no-ads.mode.ts
@@ -1,8 +1,6 @@
 import {
-	Apstag,
 	Audigent,
 	communicationService,
-	context,
 	DiProcess,
 	eventsRepository,
 	Experian,
@@ -31,9 +29,6 @@ export class NoAdsMode implements DiProcess {
 		this.removeAdSlotsPlaceholders();
 		this.noAdsDetector.addReasons(window.ads.context.opts.noAdsReasons);
 		this.dispatchJWPlayerSetupAction();
-		if (context.get('state.isLogged')) {
-			Apstag.make().init();
-		}
 
 		this.pipeline
 			.add(this.liveRampPixel, this.audigent, this.eyeota, this.liveConnect, this.experian)
