+++ src/platforms/shared/setup/base-context.setup.ts
@@ -197,7 +197,6 @@ export class BaseContextSetup implements DiProcess {
 		);
 		context.set('bidders.s2s.bidders', this.instantConfig.get('icPrebidS2sBidders', []));
 		context.set('bidders.s2s.enabled', this.instantConfig.get('icPrebidS2sBidders', []).length > 0);
-		context.set('bidders.a9.rpa', this.instantConfig.get('icA9HEM'));
 	}
 
 	private setupStickySlotContext(): void {
