+++ src/platforms/shared/tracking/tracking.setup.ts
@@ -1,6 +1,5 @@
 import {
 	adClickTracker,
-	Apstag,
 	Bidders,
 	bidderTracker,
 	communicationService,
@@ -60,19 +59,13 @@ export class TrackingSetup {
 			communicationService.on(
 				eventsRepository.IDENTITY_PARTNER_DATA_OBTAINED,
 				(eventInfo) => {
-					const { partnerName, partnerIdentityId } = eventInfo.payload;
 					this.dwTracker.track(
 						{
-							partner_name: partnerName,
-							partner_identity_id: partnerIdentityId,
+							partner_name: eventInfo.payload.partnerName,
+							partner_identity_id: eventInfo.payload.partnerIdentityId,
 						},
 						trackingUrls.IDENTITY_INFO,
 					);
-
-					if (['liveConnect', 'MediaWiki-sha256'].includes(partnerName)) {
-						const apstag = Apstag.make();
-						apstag.init().then(() => apstag.sendHEM(partnerIdentityId));
-					}
 				},
 				false,
 			);
